✨feat: 루틴 수정 구현

+++ src/main/java/com/project/likelion13th_team1/domain/routine/repository/RoutineRepository.java
@@ -3,5 +3,5 @@ package com.project.likelion13th_team1.domain.routine.repository;
 import com.project.likelion13th_team1.domain.routine.entity.Routine;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface RoutineRepository extends JpaRepository<Routine, Integer> {
+public interface RoutineRepository extends JpaRepository<Routine, Long> {
 }
